agent/requirements.txt
pydantic-ai[google]>=0.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.28.0
ag-ui-protocol>=0.1.0
copilotkit>=0.1.0
python-dotenv>=1.0.0
asyncpg>=0.30.0
